requirements.txt
pandas
tqdm
pysentimiento
joblib
pytesseract
pdf2image
Pillow
scikit-learn
xgboost
